--- a/src/Common/SLRUCachePolicy.h
+++ b/src/Common/SLRUCachePolicy.h
@@ -3,7 +3,8 @@
 #include <Common/ICachePolicy.h>
 
 #include <list>
-#include <unordered_map>
+
+#include <absl/container/flat_hash_map.h>
 
 namespace DB
 {
@@ -204,7 +205,7 @@ private:
         SLRUQueueIterator queue_iterator;
     };
 
-    using Cells = std::unordered_map<Key, Cell, HashFunction>;
+    using Cells = absl::flat_hash_map<Key, Cell, HashFunction>;
 
     Cells cells;