--- a/src/Interpreters/BloomFilter.h
+++ b/src/Interpreters/BloomFilter.h
@@ -55,6 +55,8 @@ public:
     /// For debug.
     UInt64 isEmpty() const;
 
+    size_t memoryUsageBytes() const { return filter.size() * sizeof(filter[0]); }
+
     friend bool operator== (const BloomFilter & a, const BloomFilter & b);
 private: