--- a/src/Interpreters/GinFilter.h
+++ b/src/Interpreters/GinFilter.h
@@ -77,6 +77,8 @@ public:
     const GinSegmentWithRowIdRangeVector & getFilter() const { return rowid_ranges; }
     GinSegmentWithRowIdRangeVector & getFilter() { return rowid_ranges; }
 
+    size_t memoryUsageBytes() const { return rowid_ranges.size() * sizeof(rowid_ranges[0]); }
+
 private:
     /// Filter parameters
     const GinFilterParameters & params;