--- a/src/Storages/MergeTree/MergeTreeIndexAnnoy.cpp
+++ b/src/Storages/MergeTree/MergeTreeIndexAnnoy.cpp
@@ -74,6 +74,12 @@ size_t AnnoyIndexWithSerialization<Distance>::getDimensions() const
     return Base::get_f();
 }
 
+template<typename Distance>
+size_t AnnoyIndexWithSerialization<Distance>::memoryUsageBytes() const
+{
+    return Base::_s * Base::_n_nodes;
+}
+
 
 template <typename Distance>
 MergeTreeIndexGranuleAnnoy<Distance>::MergeTreeIndexGranuleAnnoy(const String & index_name_, const Block & index_sample_block_)