--- a/src/Storages/MergeTree/MergeTreeIndexAnnoy.h
+++ b/src/Storages/MergeTree/MergeTreeIndexAnnoy.h
@@ -20,6 +20,7 @@ public:
     void serialize(WriteBuffer & ostr) const;
     void deserialize(ReadBuffer & istr);
     size_t getDimensions() const;
+    size_t memoryUsageBytes() const;
 };
 
 template <typename Distance>
@@ -38,6 +39,8 @@ struct MergeTreeIndexGranuleAnnoy final : public IMergeTreeIndexGranule
     void deserializeBinary(ReadBuffer & istr, MergeTreeIndexVersion version) override;
 
     bool empty() const override { return !index.get(); }
+
+    size_t memoryUsageBytes() const override { return index->memoryUsageBytes(); }
 
     const String index_name;
     const Block index_sample_block;