### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Added a cache for deserialized secondary index granules. This should make repeated queries that use secondary index faster, if the index for the whole table fits in the cache. Size of the new cache is controlled by server setting `secondary_index_cache_size`.

Pretty straightforward. A lot of copypasta from UncompressedCache.

Should help a lot with usearch ANN indexes, when they all fit in memory. Making it fast when the index doesn't fit in memory would take more work, maybe even changing usearch's serialization format to have more locality.