@UnamedRus 2023-08-31
ClickHouse already have skip index cache, this will be different?

https://github.com/ClickHouse/ClickHouse/pull/27961