@al13n321 2023-08-31
Hm, right, maybe we should remove the uncompressed cache for index if this new cache works out.

The idea is to avoid deserializing the index for each query. If the index allows doing the lookup faster than a full scan then deserializing the whole index in O(n) time on each query partially defeats the purpose.

(Alternatively, we could say that index should be able to work without deserialization, directly from a SeekableReadBuffer. But that would take considerable extra work to support, in particular for usearch, at least currently. But maybe we'll do it anyway, to avoid caching the whole index if we only access small parts of it.)