@al13n321 2023-11-17
Now this PR undoes https://github.com/ClickHouse/ClickHouse/pull/55683 . Seems impossible to have both: if granules are inserted into cache, they can't be reused.