@jrdi 2023-12-07
We already have cache and uncompressed cache for secondary indexes. What does this cache differently?

https://github.com/ClickHouse/ClickHouse/blob/add243593d950c46eb5ec46a780119533d9411b1/src/Core/ServerSettings.h#L68-L73