@alexey-milovidov 2023-12-27
@jrdi, the index is still deserialized, and it can take a lot of time.
The indices can be large (larger than the data) in the case of vector search (HNSW), and it makes sense to just keep the deserialized data structure in memory.

Although this scenario is limited, but vector search is often used on minuscule datasets (under one billion records).