@azat 2024-01-26
FWIW here are some numbers, shortly it gives significant boost for ANN indexes. I've tested on a table with 33'114'778 rows with embeddings (array of 768 elements):

- without index - 216 sec (4 seconds with data in page cache)
- annoy before - 100sec (it is just too suboptimal even in memory)
- usearch before - 33sec
- usearch with this patch - 0.064sec