@azat 2024-01-26
But funny thing that it actually may slow down regular indexes.

Before:

```sql

```