@azat 2024-01-26
But funny thing that it actually may slow down regular indexes.

<details>

<summary>Input</summary>

```sql
CREATE TABLE data
(
    `part` Int32,
    `key` Int32,
    `v1` Int32,
    `s` SimpleAggregateFunction(sum, Int64),
    INDEX v1_index v1 TYPE minmax GRANULARITY 1
)
ENGINE = AggregatingMergeTree
PARTITION BY part
ORDER BY key
SETTINGS index_granularity = 1;

insert into data (part, key, v1) select number%100, number, number from numbers(1e6);
```

</details>

Before:

```sql
EXPLAIN PIPELINE
SELECT *
FROM data
FINAL
WHERE v1 > 1
SETTINGS max_threads = 2, max_final_threads = 2, force_data_skipping_indices = 'v1_index', use_skip_indexes_if_final = 0;
-- 16 rows in set. Elapsed: 0.396 sec. 

EXPLAIN PIPELINE
SELECT *
FROM data
FINAL
WHERE v1 > 1
SETTINGS max_threads = 2, max_final_threads = 2, force_data_skipping_indices = 'v1_index', use_skip_indexes_if_final = 1;
-- 16 rows in set. Elapsed: 0.756 sec. 
```

After:

```sql
EXPLAIN PIPELINE
SELECT *
FROM data
FINAL
WHERE v1 > 1
SETTINGS max_threads = 2, max_final_threads = 2, force_data_skipping_indices = 'v1_index', use_skip_indexes_if_final = 0;
-- 16 rows in set. Elapsed: 0.307 sec. 

EXPLAIN PIPELINE
SELECT *
FROM data
FINAL
WHERE v1 > 1
SETTINGS max_threads = 2, max_final_threads = 2, force_data_skipping_indices = 'v1_index', use_skip_indexes_if_final = 1;
-- 16 rows in set. Elapsed: 1.496 sec. 
```

So now it is **2x slower**, perf points to `SLRUCachePolicy`

<details>

<summary>perf</summary>

Perf `use_skip_indexes_if_final=0`

```
Samples: 5K of event 'cycles:P', 4000 Hz, Event count (approx.): 5139605993 lost: 0/0 drop: 0/0
Overhead  Shared Object     Symbol
  15.11%  clickhouse        [.] DB::MergeTreeIndexGranuleMinMax::deserializeBinary(DB::ReadBuffer&, unsigned char)
   9.47%  clickhouse        [.] DB::buildPipesForReadingByPKRanges(DB::KeyDescription const&, std::__1::shared_ptr<DB::ExpressionActions>, DB::RangesInDataParts, unsigned long, std::__1::shared_ptr<DB::Context const>, std::__1::function<DB::Pipe (DB::RangesInDataParts)>&&)
   4.29%  clickhouse        [.] DB::Field::~Field()
```

Perf `use_skip_indexes_if_final=1` before

```
Samples: 2K of event 'cycles:P', 4000 Hz, Event count (approx.): 2420778913 lost: 0/0 drop: 0/523
Overhead  Shared Object     Symbol
  26.88%  clickhouse        [.] DB::buildPipesForReadingByPKRanges(DB::KeyDescription const&, std::__1::shared_ptr<DB::ExpressionActions>, DB::RangesInDataParts, unsigned long, std::__1::shared_ptr<DB::Context const>, std::__1::function<DB::Pipe (DB::RangesInDataParts)>&&)
   7.50%  clickhouse        [.] auto DB::Field::dispatch<DB::applyVisitor<DB::FieldVisitorAccurateLess, DB::Field const&, DB::Field const&>(DB::FieldVisitorAccurateLess&&, DB::Field const&, DB::Field const&)::{lambda(auto:1&)#1}::operator()<long const>(DB::FieldVisitorAccurateLess&) const::{lam
   7.04%  clickhouse        [.] auto DB::Field::dispatch<DB::applyVisitor<DB::FieldVisitorAccurateLess, DB::Field const&, DB::Field const&>(DB::FieldVisitorAccurateLess&&, DB::Field const&, DB::Field const&)::{lambda(auto:1&)#1}, DB::Field const&>(DB::FieldVisitorAccurateLess&&, DB::Field const
   6.38%  clickhouse        [.] void std::__1::__sift_up[abi:v15000]<std::__1::_ClassicAlgPolicy, std::__1::less<(anonymous namespace)::split(DB::RangesInDataParts, unsigned long)::PartsRangesIterator>&, std::__1::__wrap_iter<(anonymous namespace)::split(DB::RangesInDataParts, unsigned long)::P
   4.56%  clickhouse        [.] malloc
```

Perf `use_skip_indexes_if_final=1` after

```
Samples: 30K of event 'cycles', 4000 Hz, Event count (approx.): 12978901988 lost: 0/0 drop: 27/36                                                                                             
Overhead  Shared Object     Symbol                                                                                                                                                            
  17.11%  clickhouse        [.] DB::SLRUCachePolicy<wide::integer<128ul, unsigned int>, DB::SecondaryIndexCacheCell, UInt128TrivialHash, DB::SecondaryIndexCacheWeightFunction>::get          
  11.48%  clickhouse        [.] DB::MergeTreeIndexReader::read                                                                                                                                
   5.86%  clickhouse        [.] DB::buildPipesForReadingByPKRanges                                                                                                                            
```

</details>