@KrisAnTis-Group 2024-02-07
we really need this 515x speedup in vector search 🥹